// src/types/api.ts
/**
 * Базовый интерфейс для ответов API
 */
export interface ApiResponse<T = any> {
  /**
   * Флаг успешности запроса
   */
  success: boolean;
  
  /**
   * Сообщение для отображения пользователю
   */
  message?: string;
  
  /**
   * Сообщение об ошибке (в случае неуспешного запроса)
   */
  error?: string;
  
  /**
   * Данные, возвращаемые API (в случае успешного запроса)
   */
  data?: T;
} 